Nghia Pham

+++ src/main/java/employee/com/repository/PositionRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import employee.com.entity.PositionEntity;
 
-public interface PositionRepository extends JpaRepository<PositionEntity, Integer> {
+public interface PositionRepository extends JpaRepository<PositionEntity, Long> {
+
 	PositionEntity findByName(String name);
+
 }
